app_model/requirements.txt
tensorflow==2.15.0
keras==2.15.0
annoy==1.17.3
fastapi==0.111.0  
google-cloud-storage==2.16.0 
numpy==1.26.4    
pandas==2.2.2  
pydantic==2.7.1 
sentence-transformers==2.7.0  
uvicorn==0.29.0
